int converted to Long

+++ generator/src/main/java/org/abcd/examples/ParLang/AstNodes/IntegerNode.java
@@ -2,9 +2,9 @@ package org.abcd.examples.ParLang.AstNodes;
 
 import org.abcd.examples.ParLang.NodeVisitor;
 
-public class IntegerNode extends LiteralNode<Integer>{
-    public IntegerNode(int value){super(value);}
-    public IntegerNode(int value, boolean isNegative){
+public class IntegerNode extends LiteralNode<Long>{
+    public IntegerNode(long value){super(value);}
+    public IntegerNode(long value, boolean isNegative){
         super(value);
         this.setIsNegated(isNegative);
     }
